GitButler Integration Commit
This is an integration commit for the virtual branches that GitButler is tracking.

Due to GitButler managing multiple virtual branches, you cannot switch back and
forth between git branches and virtual branches easily. 

If you switch to another branch, GitButler will need to be reinitialized.
If you commit on this branch, GitButler will throw it away.

Here are the branches that are currently applied:
 - arm-shooter (refs/gitbutler/arm-shooter)
   branch head: 265d661d959e16675170d8e966cda47afb423726
   - src/main/java/frc/robot/commands/PickupNote.java
   - src/main/java/frc/robot/subsystems/ArmSubsystem.java
   - src/main/java/frc/robot/commands/AutoAlign.java
   - src/main/java/frc/robot/subsystems/FiringSubsystem.java
   - src/main/java/frc/robot/RobotContainer.java
   - src/main/java/frc/robot/Constants.java
   - src/main/java/frc/robot/commands/FireNote.java
   - src/main/java/frc/robot/commands/SetArmValue.java

Your previous branch was: refs/heads/arm-shooter

The sha for that commit was: 265d661d959e16675170d8e966cda47afb423726

For more information about what we're doing here, check out our docs:
https://docs.gitbutler.com/features/virtual-branches/integration-branch